Se ha refactorizado el añadido de nuevo cierre
